--- packages/edge-server/src/main/incoming/message/send.ts
@@ -1,14 +1,0 @@
-import { valueToBuffer, encodeErrorResponse } from '../../../protocol'
-import { BasedErrorCode, ErrorPayload, createError } from '../../error'
-import { WebsocketClient } from '../../../types'
-import { BasedServer } from '../../server'
-
-export function sendError<T extends BasedErrorCode>(
-  server: BasedServer,
-  client: WebsocketClient,
-  basedCode: T,
-  payload: ErrorPayload[T]
-): void {
-  const errorData = createError(server, client, basedCode, payload)
-  client.ws?.send(encodeErrorResponse(valueToBuffer(errorData)), true, false)
-}
