+++ packages/edge-server/src/main/observable/extendCache.ts
@@ -1,0 +1,8 @@
+import { ActiveObservable } from '../../types'
+
+export const extendCache = (obs: ActiveObservable) => {
+  if (obs.beingDestroyed) {
+    clearTimeout(obs.beingDestroyed)
+    obs.beingDestroyed = null
+  }
+}
