+++ packages/edge-server/src/main/observable/get.ts
@@ -1,0 +1,13 @@
+import { ActiveObservable } from '../../types'
+import { BasedServer } from '../server'
+import { extendCache } from './extendCache'
+
+export const getObs = (server: BasedServer, id: number): ActiveObservable => {
+  const obs = server.activeObservablesById.get(id)
+  extendCache(obs)
+  return obs
+}
+
+export const hasObs = (server: BasedServer, id: number): Boolean => {
+  return server.activeObservablesById.has(id)
+}
